SYNC-1814 Ensure corresponding SyncDLObject exists before deleting DLSyncEvent

+++ modules/apps/sync/sync-service/src/main/java/com/liferay/sync/internal/messaging/SyncMaintenanceMessageListener.java
@@ -17,14 +17,14 @@ package com.liferay.sync.internal.messaging;
 import com.liferay.document.library.kernel.model.DLSyncEvent;
 import com.liferay.document.library.kernel.service.DLSyncEventLocalService;
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
-import com.liferay.portal.kernel.dao.orm.DynamicQuery;
 import com.liferay.portal.kernel.dao.orm.Property;
 import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.messaging.BaseMessageListener;
+import com.liferay.portal.kernel.messaging.DestinationNames;
 import com.liferay.portal.kernel.messaging.Message;
+import com.liferay.portal.kernel.messaging.MessageBusUtil;
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.scheduler.SchedulerEngineHelper;
 import com.liferay.portal.kernel.scheduler.SchedulerEntry;
@@ -32,10 +32,16 @@ import com.liferay.portal.kernel.scheduler.SchedulerEntryImpl;
 import com.liferay.portal.kernel.scheduler.TimeUnit;
 import com.liferay.portal.kernel.scheduler.Trigger;
 import com.liferay.portal.kernel.scheduler.TriggerFactory;
+import com.liferay.portal.kernel.transaction.TransactionCommitCallbackUtil;
 import com.liferay.portal.kernel.util.Time;
+import com.liferay.sync.model.SyncDLObject;
 import com.liferay.sync.service.SyncDLFileVersionDiffLocalService;
 import com.liferay.sync.service.SyncDLObjectLocalService;
 import com.liferay.sync.service.configuration.SyncServiceConfigurationValues;
+
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
@@ -57,8 +63,10 @@ public class SyncMaintenanceMessageListener extends BaseMessageListener {
 
 		String className = clazz.getName();
 
+		Date startDate = new Date(System.currentTimeMillis() + Time.HOUR);
+
 		Trigger trigger = _triggerFactory.createTrigger(
-			className, className, null, null, 1, TimeUnit.HOUR);
+			className, className, startDate, null, 1, TimeUnit.HOUR);
 
 		SchedulerEntry schedulerEntry = new SchedulerEntryImpl(
 			className, trigger);
@@ -91,31 +99,50 @@ public class SyncMaintenanceMessageListener extends BaseMessageListener {
 				_dlSyncEventLocalService.getActionableDynamicQuery();
 
 			actionableDynamicQuery.setAddCriteriaMethod(
-				new ActionableDynamicQuery.AddCriteriaMethod() {
+				dynamicQuery -> {
+					Property modifiedTimeProperty = PropertyFactoryUtil.forName(
+						"modifiedTime");
 
-					@Override
-					public void addCriteria(DynamicQuery dynamicQuery) {
-						Property modifiedTime = PropertyFactoryUtil.forName(
-							"modifiedTime");
-
-						long latestModifiedTime =
-							_syncDLObjectLocalService.getLatestModifiedTime();
-
-						dynamicQuery.add(
-							modifiedTime.le(latestModifiedTime - Time.HOUR));
-					}
-
+					dynamicQuery.add(
+						modifiedTimeProperty.le(
+							_syncDLObjectLocalService.getLatestModifiedTime()));
 				});
 			actionableDynamicQuery.setPerformActionMethod(
-				new ActionableDynamicQuery.PerformActionMethod<DLSyncEvent>() {
+				(DLSyncEvent dlSyncEvent) -> {
+					SyncDLObject syncDLObject =
+						_syncDLObjectLocalService.fetchSyncDLObject(
+							dlSyncEvent.getType(), dlSyncEvent.getTypePK());
 
-					@Override
-					public void performAction(DLSyncEvent dlSyncEvent)
-						throws PortalException {
+					if ((syncDLObject == null) ||
+						(dlSyncEvent.getModifiedTime() >
+							syncDLObject.getModifiedTime())) {
 
+						TransactionCommitCallbackUtil.registerCallback(
+							() -> {
+								Message dlSyncEventMessage = new Message();
+
+								Map<String, Object> values = new HashMap<>();
+
+								values.put("event", dlSyncEvent.getEvent());
+								values.put(
+									"modifiedTime",
+									dlSyncEvent.getModifiedTime());
+								values.put("type", dlSyncEvent.getType());
+								values.put("typePK", dlSyncEvent.getTypePK());
+
+								dlSyncEventMessage.setValues(values);
+
+								MessageBusUtil.sendMessage(
+									DestinationNames.
+										DOCUMENT_LIBRARY_SYNC_EVENT_PROCESSOR,
+									dlSyncEventMessage);
+
+								return null;
+							});
+					}
+					else {
 						_dlSyncEventLocalService.deleteDLSyncEvent(dlSyncEvent);
 					}
-
 				});
 
 			actionableDynamicQuery.performActions();
